Issue #12: cherry-pick all refactoring scratches to the new branch
