+++ scripts/autoscaling/azure/terminate_node.py
@@ -12,90 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import functools
-import os
-
 import pykube
 import argparse
-from azure.common.client_factory import get_client_from_auth_file
-from azure.mgmt.resource import ResourceManagementClient
-from azure.mgmt.compute import ComputeManagementClient
-
-RUN_ID_LABEL = 'runid'
-CLOUD_REGION_LABEL = 'cloud_region'
-KUBE_CONFIG_PATH = '~/.kube/config'
-resource_group_name = os.environ["AZURE_RESOURCE_GROUP"]
-
-res_client = get_client_from_auth_file(ResourceManagementClient)
-compute_client = get_client_from_auth_file(ComputeManagementClient)
+from pipeline.autoscaling import azureprovider, kubeprovider
 
 
-def resolve_azure_api(resource):
-    """ This method retrieves the latest non-preview api version for
-    the given resource (unless the preview version is the only available
-    api version) """
-    provider = res_client.providers.get(resource.id.split('/')[6])
-    rt = next((t for t in provider.resource_types
-               if t.resource_type == '/'.join(resource.type.split('/')[1:])), None)
-    if rt and 'api_versions' in rt.__dict__:
-        api_version = [v for v in rt.__dict__['api_versions'] if 'preview' not in v.lower()]
-        return api_version[0] if api_version else rt.__dict__['api_versions'][0]
-
-
-def azure_resource_type_cmp(r1, r2):
-    if str(r1.type).split('/')[-1] == "virtualMachines":
-        return -1
-    elif str(r1.type).split('/')[-1] == "networkInterfaces" and str(r2.type).split('/')[-1] != "virtualMachines":
-        return -1
-    return 0
-
-
-def delete_cloud_node(node_name):
-    vm_info = compute_client.virtual_machines.get(resource_group_name, node_name)
-    if vm_info is not None and "Name" in vm_info.tags:
-        resources = []
-        for resource in res_client.resources.list(filter="tagName eq 'Name' and tagValue eq '" + vm_info.tags["Name"] + "'"):
-            resources.append(resource)
-            # we need to sort resources to be sure that vm and nic will be deleted first, because it has attached resorces(disks and ip)
-            resources.sort(key=functools.cmp_to_key(azure_resource_type_cmp))
-        for resource in resources:
-            res_client.resources.delete(
-                resource_group_name=resource.id.split('/')[4],
-                resource_provider_namespace=resource.id.split('/')[6],
-                parent_resource_path='',
-                resource_type=str(resource.type).split('/')[-1],
-                resource_name=resource.name,
-                api_version=resolve_azure_api(resource),
-                parameters=resource
-            ).wait()
-
-
-def delete_kubernetes_node(kube_api, node_name):
-    if node_name is not None and get_node(kube_api, node_name) is not None:
-        obj = {
-            "apiVersion": "v1",
-            "kind": "Node",
-            "metadata": {
-                "name": node_name
-            }
-        }
-        pykube.Node(kube_api, obj).delete()
-
-
-def get_node(kube_api, nodename):
-    nodes = pykube.Node.objects(kube_api).filter(field_selector={'metadata.name': nodename})
-    if len(nodes.response['items']) == 0:
-        return None
-    return nodes.response['items'][0]
-
-
-def get_kube_api():
-    try:
-        api = pykube.HTTPClient(pykube.KubeConfig.from_service_account())
-    except Exception as e:
-        api = pykube.HTTPClient(pykube.KubeConfig.from_file(KUBE_CONFIG_PATH))
-    api.session.verify = False
-    return api
+def delete_kubernetes_node(kube_provider, node_name):
+    if node_name is not None and kube_provider.get_node(node_name) is not None:
+        kube_provider.delete_kubernetes_node_by_name(node_name)
 
 
 def main():
@@ -103,10 +27,15 @@ def main():
     parser.add_argument("--internal_ip", "-ip", type=str, required=True)
     parser.add_argument("--node_name", "-n", type=str, required=True)
     args = parser.parse_args()
-    kube_api = get_kube_api()
-    delete_kubernetes_node(kube_api, args.node_name)
 
-    delete_cloud_node(args.node_name)
+    kube_provider = kubeprovider.KubeProvider()
+    cloud_region = kube_provider.get_cloud_region_by_node_name(args.node_name)
+    cloud_provider = azureprovider.AzureInstanceProvider(cloud_region)
+
+    run_id = kube_provider.get_run_id_by_node_name(args.node_name)
+
+    delete_kubernetes_node(kube_provider, args.node_name)
+    cloud_provider.delete_all_by_run_id(run_id)
 
 
 if __name__ == '__main__':
